requirements.txt
flask
pandas
scikit-learn
joblib
pandas-datareader
mysql-connector-python
